requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
anyio==4.4.0
attrs==25.1.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.0
dnspython==2.6.1
docker==7.1.0
fastcore==1.7.29
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.1.0
idna==3.7
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
motor==3.5.1
multidict==6.1.0
numpy==2.2.2
optuna==4.2.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pymongo==4.8.0
pyvespa
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
